+++ src/main/java/com/mobble/mobbleserver/account/auth/dto/request/SignUpRequestDto.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.account.auth.dto.request;
 import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
 import com.mobble.mobbleserver.domain.member.Gender;
 import com.mobble.mobbleserver.domain.member.Member;
-import com.mobble.mobbleserver.refactor.ground.entity.Ground;
+import com.mobble.mobbleserver.domain.ground.Ground;
 import jakarta.validation.constraints.*;
 
 public record SignUpRequestDto(
