+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/club/core/dto/response/ClubSummaryDto.java
@@ -2,7 +2,7 @@ package com.mobble.mobbleserver.infrastructure.web.club.core.dto.response;
 
 import com.mobble.mobbleserver.domain.club.core.Club;
 import com.mobble.mobbleserver.infrastructure.persistence.club.core.projection.ClubLikeInfoDto;
-import com.mobble.mobbleserver.refactor.ground.entity.Ground;
+import com.mobble.mobbleserver.domain.ground.Ground;
 
 import java.util.List;
