+++ src/main/java/com/mobble/mobbleserver/refactor/club/clubGround/repository/ClubGroundRepository.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.refactor.club.clubGround.repository;
 
 import com.mobble.mobbleserver.refactor.club.clubGround.entity.ClubGround;
-import com.mobble.mobbleserver.refactor.ground.entity.Ground;
+import com.mobble.mobbleserver.domain.ground.Ground;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
